Move tensor-related UI code from `image.rs` to `tensor.rs`
